Move functions to tests from vmc_utils.py
Since most of them are test-related, we leave only `create_vmc_tx` and
`get_vmc_json` in vmc_utils.py, which specify the vmc tx and the
contract path, and move the others to `test_vmc_handler.py`
